+++ gluecodium/src/main/java/com/here/gluecodium/generator/cpp/CppNameCache.kt
@@ -28,7 +28,6 @@ import com.here.gluecodium.model.lime.LimeExternalDescriptor.Companion.SETTER_NA
 import com.here.gluecodium.model.lime.LimeField
 import com.here.gluecodium.model.lime.LimeNamedElement
 import com.here.gluecodium.model.lime.LimeProperty
-import com.here.gluecodium.model.lime.LimeTypesCollection
 import java.util.HashMap
 
 internal class CppNameCache(
@@ -81,14 +80,8 @@ internal class CppNameCache(
     }
 
     private fun resolveNames(limeElement: LimeNamedElement): NamesCacheEntry {
-        var parentPath = limeElement.path.parent
-        var parentElement = limeReferenceMap[parentPath.toString()] as? LimeNamedElement
-        if (parentElement is LimeTypesCollection) {
-            // A type collection doesn't correspond to any named entity in C++ generated code.
-            // So skip it and use the parent namespace instead.
-            parentPath = parentPath.parent
-            parentElement = null
-        }
+        val parentPath = limeElement.path.parent
+        val parentElement = limeReferenceMap[parentPath.toString()] as? LimeNamedElement
 
         val parentFullName: String
         var parentIsExternal = false